+++ pkg/controller/iam/iamclient/dcliamclient_test.go
@@ -19,7 +19,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/apis/iam/v1beta1"
+	"github.com/GoogleCloudPlatform/k8s-config-connector/apis/iam/v1beta1"
 
 	dcliam "github.com/GoogleCloudPlatform/declarative-resource-client-library/services/google/iam"
 	dclunstruct "github.com/GoogleCloudPlatform/declarative-resource-client-library/unstructured"
