+++ pkg/controller/iam/iamclient/externalonly.go
@@ -17,8 +17,8 @@ package iamclient
 import (
 	"fmt"
 
+	iamv1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/apis/iam/v1beta1"
 	corekccv1alpha1 "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/apis/core/v1alpha1"
-	iamv1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/apis/iam/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/gvks/externalonlygvks"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
