+++ pkg/controller/iam/partialpolicy/bindings_test.go
@@ -19,7 +19,7 @@ import (
 	"reflect"
 	"testing"
 
-	iamv1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/apis/iam/v1beta1"
+	iamv1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/apis/iam/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/iam/partialpolicy"
 
 	"github.com/google/go-cmp/cmp"
